Fixed Level 5 and added text ui testing increment
