+++ text-ui-test/EXPECTED.TXT
@@ -1,7 +1,23 @@
-Hello from
- ____        _        
-|  _ \ _   _| | _____ 
-| | | | | | | |/ / _ \
-| |_| | |_| |   <  __/
-|____/ \__,_|_|\_\___|
+Greetings, I am Nibbalim!
+Your wish is my command!
+
+Roger. This task has been added:
+  [T][ ] todo return some books
+Now you have 1 tasks in your list.
+Roger. This task has been added:
+  [D][ ] deadline complete assignment (by: Monday 2359)
+Now you have 2 tasks in your list.
+Roger. This task has been added:
+  [E][ ] event CS2103T lecture (from: Monday 1000 to: Monday 1200)
+Now you have 3 tasks in your list.
+Solid! I'm marking this task as done:
+[T][X] todo return some books
+Here are the tasks in your list:
+1.[T][X] todo return some books
+2.[D][ ] deadline complete assignment (by: Monday 2359)
+3.[E][ ] event CS2103T lecture (from: Monday 1000 to: Monday 1200)
+Aight, marking this as not done:
+[T][ ] todo return some books
+DukeException: TODO needs a description!
+Kay thanks bye!
